+++ src/cpp/task_manager_imp.hpp
@@ -26,7 +26,7 @@ public:
     
     static std::shared_ptr<TaskManagerGen> instance();
     
-    virtual void addTask(int32_t task_id, int64_t delay, int64_t repeated, const std::shared_ptr<TaskExcuserGen> & task);
+    virtual void addTask(int64_t task_id, int64_t delay, int64_t repeated, const std::shared_ptr<TaskExcuserGen> & task);
     virtual void addTaskInfo(const TaskInfo & task, const std::shared_ptr<TaskExcuserGen> & task_excuser);
     virtual void addTaskExcuser(const std::shared_ptr<TaskExcuserGen> & task);
